feat: Add `boosters.json` and `guild.json` schemas
