+++ packages/schemas/src/schemas/hypixel/guild.json
@@ -1,0 +1,289 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://starboard.ecr.dev/schemas/hypixel/HypixelGuildResponse.json",
+  "title": "HypixelGuildResponse",
+  "description": "Schema for data returned from https://api.hypixel.net/guild",
+  "oneOf": [
+    {
+      "type": "object",
+      "properties": {
+        "success": {
+          "const": true
+        },
+        "guild": {
+          "type": "array",
+          "items": {
+            "$ref": "#/definitions/HypixelGuild"
+          }
+        }
+      },
+      "required": [
+        "success",
+        "guild"
+      ]
+    },
+    {
+      "type": "object",
+      "properties": {
+        "success": {
+          "const": false
+        },
+        "cause": {
+          "type": "string"
+        },
+        "throttle": {
+          "type": "boolean"
+        },
+        "global": {
+          "type": "boolean"
+        }
+      },
+      "required": [
+        "success",
+        "cause"
+      ]
+    }
+  ],
+  "definitions": {
+    "HypixelGuild": {
+      "type": "object",
+      "properties": {
+        "_id": {
+          "type": "string",
+          "pattern": "^[0-9a-f]{24}$"
+        },
+        "coins": {
+          "type": "integer"
+        },
+        "coinsEver": {
+          "type": "integer"
+        },
+        "created": {
+          "type": "integer"
+        },
+        "joinable": {
+          "type": "boolean"
+        },
+        "members": {
+          "type": "array",
+          "items": {
+            "type": "object",
+            "properties": {
+              "joined": {
+                "type": "integer"
+              },
+              "name": {
+                "type": "string"
+              },
+              "rank": {
+                "type": "string"
+              },
+              "uuid": {
+                "type": "string"
+              },
+              "questParticipation": {
+                "type": "integer"
+              },
+              "mutedTill": {
+                "type": "integer"
+              },
+              "expHistory": {
+                "type": "object",
+                "patternProperties": {
+                  "\\d{4}-\\d{2}-\\d{2}": {
+                    "type": "integer"
+                  }
+                }
+              }
+            }
+          }
+        },
+        "name": {
+          "type": "string"
+        },
+        "publiclyListed": {
+          "type": "boolean"
+        },
+        "banner": {
+          "type": "object",
+          "properties": {
+            "Base": {
+              "type": "integer"
+            },
+            "Patterns": {
+              "type": "array",
+              "items": {
+                "type": "object",
+                "properties": {
+                  "Pattern": {
+                    "type": "string"
+                  },
+                  "Color": {
+                    "type": "string"
+                  }
+                }
+              }
+            }
+          }
+        },
+        "tagColor": {
+          "type": "string"
+        },
+        "achievements": {
+          "type": "object",
+          "properties": {
+            "WINNERS": {
+              "type": "integer"
+            },
+            "EXPERIENCE_KINGS": {
+              "type": "integer"
+            },
+            "ONLINE_PLAYERS": {
+              "type": "integer"
+            }
+          }
+        },
+        "exp": {
+          "type": "integer"
+        },
+        "legacyRanking": {
+          "type": "integer"
+        },
+        "ranks": {
+          "type": "array",
+          "items": {
+            "type": "object",
+            "properties": {
+              "name": {
+                "type": "string"
+              },
+              "default": {
+                "type": "boolean"
+              },
+              "tag": {
+                "type": "string"
+              },
+              "created": {
+                "type": "integer"
+              },
+              "priority": {
+                "type": "integer"
+              }
+            }
+          }
+        },
+        "chatMute": {
+          "type": "integer"
+        },
+        "preferredGames": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          }
+        },
+        "description": {
+          "type": "string"
+        },
+        "name_lower": {
+          "type": "string"
+        },
+        "tag": {
+          "type": "string"
+        },
+        "guildExpByGameType": {
+          "type": "object",
+          "properties": {
+            "SMP": {
+              "type": "integer"
+            },
+            "ARENA": {
+              "type": "integer"
+            },
+            "SPEED_UHC": {
+              "type": "integer"
+            },
+            "LEGACY": {
+              "type": "integer"
+            },
+            "DUELS": {
+              "type": "integer"
+            },
+            "WALLS": {
+              "type": "integer"
+            },
+            "UHC": {
+              "type": "integer"
+            },
+            "REPLAY": {
+              "type": "integer"
+            },
+            "PIT": {
+              "type": "integer"
+            },
+            "MURDER_MYSTERY": {
+              "type": "integer"
+            },
+            "BEDWARS": {
+              "type": "integer"
+            },
+            "TNTGAMES": {
+              "type": "integer"
+            },
+            "SURVIVAL_GAMES": {
+              "type": "integer"
+            },
+            "SUPER_SMASH": {
+              "type": "integer"
+            },
+            "BUILD_BATTLE": {
+              "type": "integer"
+            },
+            "WALLS3": {
+              "type": "integer"
+            },
+            "SKYWARS": {
+              "type": "integer"
+            },
+            "HOUSING": {
+              "type": "integer"
+            },
+            "PROTOTYPE": {
+              "type": "integer"
+            },
+            "ARCADE": {
+              "type": "integer"
+            },
+            "SKYBLOCK": {
+              "type": "integer"
+            },
+            "GINGERBREAD": {
+              "type": "integer"
+            },
+            "VAMPIREZ": {
+              "type": "integer"
+            },
+            "WOOL_GAMES": {
+              "type": "integer"
+            },
+            "PAINTBALL": {
+              "type": "integer"
+            },
+            "QUAKECRAFT": {
+              "type": "integer"
+            },
+            "BATTLEGROUND": {
+              "type": "integer"
+            },
+            "MCGO": {
+              "type": "integer"
+            }
+          }
+        }
+      },
+      "required": [
+        "_id",
+        "name"
+      ]
+    }
+  }
+}
